+++ src/dom/special_attributes.ts
@@ -1,0 +1,11 @@
+export interface ValueableHTMLElement extends HTMLElement {
+  [SpecialAttributes.Value]: string;
+}
+
+/**
+ * Attributes and properties that must be updated by library
+ * in order to show reactivity.
+ */
+export const enum SpecialAttributes {
+  Value = 'value',
+}