+++ src/utils/regexps.ts
@@ -6,4 +6,8 @@ export const booleanAttributeRegExp = /\?([\w-]+)=$/;
 
 export const hookAttributeRegExp = /:(mounted|updated|removed)=$/;
 
+export const dataHookIdJSRegExp = /hookId(Mounted|Updated|Removed)-[\d]{1,}$/;
+
 export const pathVariableRegExp = /:.+:(\?)?/g;
+
+export const valueRegExp = /(value)=$/;